# api-specification.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Organix Backend API",
    "description": "API RESTful para sistema de gestão de conteúdo multi-tenant",
    "contact": { "name": "Organix Team", "email": "contact@organix.com" },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.0.0"
  },
  "servers": [
    { "url": "http://localhost:8080", "description": "Generated server url" }
  ],
  "security": [{ "bearerAuth": [] }],
...
  "paths": {
...
    "/api/company": {
      "get": {
        "tags": ["Company Management"],
        "summary": "Buscar informações da empresa",
        "description": "Retorna as informações da empresa do usuário autenticado",
        "operationId": "getCompanyInfo",
        "responses": {
          "404": {
            "description": "Empresa não encontrada",
            "content": {
              "*/*": {
                "schema": { "$ref": "#/components/schemas/CompanyResponse" }
              }
            }
          },
          "200": {
            "description": "Informações da empresa retornadas com sucesso",
            "content": {
              "*/*": {
                "schema": { "$ref": "#/components/schemas/CompanyResponse" }
              }
            }
          }
        },
        "security": [{ "bearerAuth": [] }]
      },
      "put": {
        "tags": ["Company Management"],
        "summary": "Atualizar informações da empresa",
        "description": "Atualiza as informações da empresa (apenas administradores)",
        "operationId": "updateCompany",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateCompanyRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Empresa atualizada com sucesso",
            "content": {
              "*/*": {
                "schema": { "$ref": "#/components/schemas/CompanyResponse" }
              }
            }
          },
          "404": {
            "description": "Empresa não encontrada",
            "content": {
              "*/*": {
                "schema": { "$ref": "#/components/schemas/CompanyResponse" }
              }
            }
          },
          "400": {
            "description": "Dados inválidos",
            "content": {
              "*/*": {
                "schema": { "$ref": "#/components/schemas/CompanyResponse" }
              }
            }
          },
          "403": {
            "description": "Acesso negado - apenas administradores podem atualizar a empresa",
            "content": {
              "*/*": {
                "schema": { "$ref": "#/components/schemas/CompanyResponse" }
              }
            }
          }
        },
        "security": [{ "bearerAuth": [] }]
      }
    },
...
    "/api/auth/signup": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Registro",
        "description": "Registra uma nova empresa e usuário administrador",
        "operationId": "signup",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SignupRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Dados inválidos ou email já existe",
            "content": {
              "*/*": {
                "schema": { "$ref": "#/components/schemas/JwtResponse" }
              }
            }
          },
          "200": {
            "description": "Registro realizado com sucesso",
            "content": {
              "*/*": {
                "schema": { "$ref": "#/components/schemas/JwtResponse" }
              }
            }
          }
        }
      }
    },
    "/api/auth/refresh": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Refresh Token",
        "description": "Renova o token JWT",
        "operationId": "refresh",
        "responses": {
          "401": {
            "description": "Token inválido ou expirado",
            "content": {
              "*/*": {
                "schema": { "$ref": "#/components/schemas/JwtResponse" }
              }
            }
          },
          "200": {
            "description": "Token renovado com sucesso",
            "content": {
              "*/*": {
                "schema": { "$ref": "#/components/schemas/JwtResponse" }
              }
            }
          }
        }
      }
    },
...
    "/api/auth/login": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Login",
        "description": "Autentica um usuário e retorna o token JWT",
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Login realizado com sucesso",
            "content": {
              "*/*": {
                "schema": { "$ref": "#/components/schemas/JwtResponse" }
              }
            }
          },
          "401": {
            "description": "Credenciais inválidas",
            "content": {
              "*/*": {
                "schema": { "$ref": "#/components/schemas/JwtResponse" }
              }
            }
          }
        }
      }
    },
...
    "/api/company/stats": {
      "get": {
        "tags": ["Company Management"],
        "summary": "Estatísticas da empresa",
        "description": "Retorna estatísticas gerais da empresa",
        "operationId": "getCompanyStats",
        "responses": {
          "200": {
            "description": "Estatísticas retornadas com sucesso",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyStatsResponse"
                }
              }
            }
          }
        },
        "security": [{ "bearerAuth": [] }]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "UpdateCompanyRequest": {
        "required": ["name"],
        "type": "object",
        "properties": {
          "name": { "maxLength": 100, "minLength": 0, "type": "string" },
          "industry": { "maxLength": 50, "minLength": 0, "type": "string" },
          "size": { "maxLength": 20, "minLength": 0, "type": "string" },
          "website": { "maxLength": 255, "minLength": 0, "type": "string" },
          "description": { "maxLength": 500, "minLength": 0, "type": "string" }
        },
        "description": "Novos dados da empresa"
      },
      "CompanyResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "industry": { "type": "string" },
          "size": { "type": "string" },
          "website": { "type": "string" },
          "description": { "type": "string" },
          "adminId": { "type": "string", "format": "uuid" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "SignupRequest": {
        "required": ["adminName", "companyName", "email", "password"],
        "type": "object",
        "properties": {
          "companyName": { "maxLength": 100, "minLength": 0, "type": "string" },
          "adminName": { "maxLength": 100, "minLength": 0, "type": "string" },
          "email": { "type": "string" },
          "password": {
            "maxLength": 2147483647,
            "minLength": 6,
            "type": "string"
          }
        },
        "description": "Dados para registro"
      },
      "CompanyInfoDto": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" },
          "adminId": { "type": "string", "format": "uuid" }
        }
      },
      "JwtResponse": {
        "type": "object",
        "properties": {
          "user": { "$ref": "#/components/schemas/UserInfoDto" },
          "company": { "$ref": "#/components/schemas/CompanyInfoDto" },
          "token": { "type": "string" },
          "refreshToken": { "type": "string" }
        }
      },
      "UserInfoDto": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "email": { "type": "string" },
          "adminType": { "type": "string", "enum": ["ADMIN", "OPERATOR"] },
          "companyId": { "type": "string", "format": "uuid" }
        }
      },
      "LoginRequest": {
        "required": ["email", "password"],
        "type": "object",
        "properties": {
          "email": { "type": "string" },
          "password": { "type": "string" }
        },
        "description": "Credenciais de login"
      },
...
      "CompanyStatsResponse": {
        "type": "object",
        "properties": {
          "companyId": { "type": "string", "format": "uuid" },
          "companyName": { "type": "string" },
          "totalUsers": { "type": "integer", "format": "int64" },
          "totalProducts": { "type": "integer", "format": "int64" },
          "totalDrafts": { "type": "integer", "format": "int64" },
          "pendingDrafts": { "type": "integer", "format": "int64" },
          "approvedDrafts": { "type": "integer", "format": "int64" },
          "totalContent": { "type": "integer", "format": "int64" },
          "publishedContent": { "type": "integer", "format": "int64" },
          "scheduledContent": { "type": "integer", "format": "int64" }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "JWT Token para autenticação. Formato: Bearer {token}",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}